chore: Update npm dependencies
This commit updates the npm dependencies in the `package.json` file. It adds the following packages:
- `@radix-ui/react-tooltip` version 1.0.7
- `class-variance-authority` version 0.7.0
- `clsx` version 2.1.1
- `tailwind-merge` version 2.3.0
- `tailwindcss-animate` version 1.0.7

It also updates the `tailwindcss` package to version 3.4.4.

The purpose of this update is to ensure that the project is using the latest stable versions of the dependencies and to take advantage of any bug fixes or new features introduced in these versions.