chore: Update npm dependencies and tailwind.config.js
This commit updates the npm dependencies in the `package.json` file and the `tailwind.config.js` file.

In `package.json`, the following packages were added:
- `@radix-ui/react-tooltip` version 1.0.7
- `class-variance-authority` version 0.7.0
- `clsx` version 2.1.1
- `tailwind-merge` version 2.3.0
- `tailwindcss-animate` version 1.0.7

The `tailwindcss` package was also updated to version 3.4.4.

In `tailwind.config.js`, the `tailwindcss-animate` plugin was imported using the ES6 import syntax.

The purpose of these changes is to ensure that the project is using the latest stable versions of the dependencies and to take advantage of any bug fixes or new features introduced in these versions. Additionally, the update to `tailwind.config.js` improves the organization and readability of the code.