chore: Update npm dependencies to latest versions
This commit updates the npm dependencies in the `package.json` file to their latest stable versions. The following packages were updated:
- `@radix-ui/react-tooltip` to version 1.0.7
- `class-variance-authority` to version 0.7.0
- `clsx` to version 2.1.1
- `tailwind-merge` to version 2.3.0
- `tailwindcss-animate` to version 1.0.7

Additionally, the `tailwindcss` package was updated to version 3.4.4.

The purpose of this update is to ensure that the project is using the most recent versions of the dependencies, benefiting from bug fixes and new features introduced in these versions.