chore: Remove unused components and update dependencies
This commit removes the unused `PlaceholderUpdater` component and the `SearchOrderInput` component. These components are no longer necessary and their removal improves code organization and reduces clutter.

Additionally, the `package.json` file is updated to include the `@radix-ui/react-alert-dialog` package version 1.1.1. This ensures that the project is using the latest stable version of the package.

The purpose of these changes is to remove unnecessary code and dependencies, improving code maintainability and reducing potential issues.